requirements.txt
praw==7.7.1
flask==3.0.3
apscheduler==3.10.4
anthropic>=0.39.0
python-dotenv==1.0.1
peewee==3.17.6
